Break predictors.py into separate files
